feat: 钱包操作

+++ common/src/main/java/com/zj/auction/common/model/Wallet.java
@@ -21,7 +21,7 @@ public class Wallet implements Serializable {
     /**
      * 基金类型
      */
-    private Integer fundType;
+    private Byte fundType;
 
     /**
      * 资金
